+++ languages/Cpp/m3_hron/m3_hron.vcxproj
@@ -150,16 +150,16 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="hron_parser.h" />
-    <ClInclude Include="hron_parser.hpp" />
+    <ClInclude Include="hron_parser\hron_parser.h" />
+    <ClInclude Include="hron_parser\hron_parser.hpp" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="targetver.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="hron_parser.c">
+    <ClCompile Include="hron_parser\hron_parser.c">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">NotUsing</PrecompiledHeader>
-      <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">NotUsing</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">NotUsing</PrecompiledHeader>
+      <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">NotUsing</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">NotUsing</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="m3_hron.cpp" />
@@ -171,7 +171,8 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <None Include="generated_hron_parser.i" />
+    <None Include="hron_parser\generated_hron_parser.i" />
+    <None Include="hron_parser\generated_hron_parser.tt" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
